+++ src/main/java/nextstep/domain/Role.java
@@ -1,0 +1,9 @@
+package nextstep.domain;
+
+public enum Role {
+    ADMIN("admin"),
+    USER("user");
+
+    Role(String role) {
+    }
+}
