+++ src/main/java/nextstep/domain/Theme.java
@@ -1,5 +1,7 @@
 package nextstep.domain;
 
+import nextstep.dto.request.ThemeRequest;
+
 public class Theme {
     private Long id;
     private String name;
@@ -20,6 +22,12 @@ public class Theme {
         this.name = name;
         this.desc = desc;
         this.price = price;
+    }
+
+    public Theme(ThemeRequest themeRequest) {
+        this.name = themeRequest.getName();
+        this.desc = themeRequest.getDesc();
+        this.price = themeRequest.getPrice();
     }
 
     public Long getId() {
