+++ src/main/java/nextstep/dto/request/ScheduleRequest.java
@@ -1,17 +1,12 @@
 package nextstep.dto.request;
 
-import nextstep.domain.Schedule;
-import nextstep.domain.Theme;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
 public class ScheduleRequest {
     private Long themeId;
     private String date;
     private String time;
 
-    public ScheduleRequest() {}
+    public ScheduleRequest() {
+    }
 
     public ScheduleRequest(Long themeId, String date, String time) {
         this.themeId = themeId;
@@ -29,13 +24,5 @@ public class ScheduleRequest {
 
     public String getTime() {
         return time;
-    }
-
-    public Schedule toEntity(Theme theme) {
-        return new Schedule(
-                theme,
-                LocalDate.parse(this.date),
-                LocalTime.parse(this.time)
-        );
     }
 }
