+++ src/main/java/nextstep/service/MemberService.java
@@ -2,19 +2,18 @@ package nextstep.service;
 
 import nextstep.dao.MemberDao;
 import nextstep.domain.Member;
-import nextstep.dto.request.MemberRequest;
 import org.springframework.stereotype.Service;
 
 @Service
 public class MemberService {
-    private MemberDao memberDao;
+    private final MemberDao memberDao;
 
     public MemberService(MemberDao memberDao) {
         this.memberDao = memberDao;
     }
 
-    public Long create(MemberRequest memberRequest) {
-        return memberDao.save(memberRequest.toEntity());
+    public Long create(Member member) {
+        return memberDao.save(member);
     }
 
     public Member findByMemberId(Long id) {
