+++ src/main/java/nextstep/service/ThemeService.java
@@ -2,7 +2,6 @@ package nextstep.service;
 
 import nextstep.dao.ThemeDao;
 import nextstep.domain.Theme;
-import nextstep.dto.request.ThemeRequest;
 import nextstep.support.exception.NotExistEntityException;
 import org.springframework.stereotype.Service;
 
@@ -10,21 +9,21 @@ import java.util.List;
 
 @Service
 public class ThemeService {
-    private ThemeDao themeDao;
+    private final ThemeDao themeDao;
 
     public ThemeService(ThemeDao themeDao) {
         this.themeDao = themeDao;
     }
 
-    public Long create(ThemeRequest themeRequest) {
-        return themeDao.save(themeRequest.toEntity());
+    public Long create(Theme theme) {
+        return themeDao.save(theme);
     }
 
     public List<Theme> findAll() {
         return themeDao.findAll();
     }
 
-    public void delete(Long id) {
+    public void delete(long id) {
         Theme theme = themeDao.findById(id);
         if (theme == null) {
             throw new NotExistEntityException();
