+++ src/test/java/nextstep/controller/MemberE2ETest.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.controller;
 
 import io.restassured.RestAssured;
 import nextstep.dto.request.MemberRequest;
@@ -12,18 +12,20 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.core.Is.is;
 
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-public class AuthE2ETest {
+public class MemberE2ETest {
     public static final String USERNAME = "username";
     public static final String PASSWORD = "password";
-    private Long memberId;
+    public static final String NAME = "name";
+    public static final String PHONE = "010-1234-5678";
+    public Long memberId;
 
     @BeforeEach
     void setUp() {
-        MemberRequest body = new MemberRequest(USERNAME, PASSWORD, "name", "010-1234-5678");
+        MemberRequest body = new MemberRequest(USERNAME, PASSWORD, NAME, PHONE);
         var memberResponse = RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -37,19 +39,43 @@ public class AuthE2ETest {
         memberId = Long.parseLong(memberLocation[memberLocation.length - 1]);
     }
 
-    @DisplayName("토큰을 생성한다")
+    @DisplayName("멤버를 생성한다")
     @Test
     public void create() {
+        MemberRequest body = new MemberRequest("A", "B", "C", "010-1111-2222");
+        RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(body)
+                .when().post("/members")
+                .then().log().all()
+                .statusCode(HttpStatus.CREATED.value());
+    }
+
+    @DisplayName("내 정보를 확인한다")
+    @Test
+    public void memberMe() {
+        // TODO: USERNAME to memberId
         TokenRequest body = new TokenRequest(memberId, PASSWORD);
-        var response = RestAssured
+        TokenResponse tokenResponse = RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(body)
                 .when().post("/login/token")
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value())
-                .extract();
+                .extract().response().as(TokenResponse.class);
 
-        assertThat(response.as(TokenResponse.class)).isNotNull();
+        RestAssured
+                .given().log().all()
+                .header("Authorization", tokenResponse.getAccessToken())
+                .when().get("/members/me")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body("username", is(USERNAME))
+                .body("password", is(PASSWORD))
+                .body("name", is(NAME))
+                .body("phone", is(PHONE));
     }
 }
