+++ src/test/java/nextstep/controller/ScheduleE2ETest.java
@@ -1,6 +1,7 @@
-package nextstep.admin;
+package nextstep.controller;
 
 import io.restassured.RestAssured;
+import nextstep.dto.request.ScheduleRequest;
 import nextstep.dto.request.ThemeRequest;
 import nextstep.dto.request.TokenRequest;
 import nextstep.dto.response.TokenResponse;
@@ -14,94 +15,95 @@ import org.springframework.test.annotation.DirtiesContext;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-public class AdminE2ETest {
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class ScheduleE2ETest {
     private static final long ADMIN_MEMBER_ID = 1L;
     private static final String ADMIN_PASSWORD = "admin";
 
+    private Long themeId;
     private String adminAccessToken;
 
     @BeforeEach
     void setUp() {
-        TokenRequest tokenRequest = new TokenRequest(ADMIN_MEMBER_ID, ADMIN_PASSWORD);
+        TokenRequest adminTokenRequest = new TokenRequest(ADMIN_MEMBER_ID, ADMIN_PASSWORD);
         this.adminAccessToken = RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(tokenRequest)
+                .body(adminTokenRequest)
                 .when().post("/login/token")
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value())
                 .extract().response().as(TokenResponse.class).getAccessToken();
-    }
 
-    @DisplayName("관리자는 테마를 생성할 수 있다")
-    @Test
-    public void createByAdmin() {
-        ThemeRequest body = new ThemeRequest("테마이름", "테마설명", 22000);
-        RestAssured
+        ThemeRequest themeRequest = new ThemeRequest("테마이름", "테마설명", 22000);
+        var response = RestAssured
                 .given().log().all()
                 .header("Authorization", this.adminAccessToken)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(body)
+                .body(themeRequest)
                 .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(HttpStatus.CREATED.value())
+                .extract();
+        String[] themeLocation = response.header("Location").split("/");
+        themeId = Long.parseLong(themeLocation[themeLocation.length - 1]);
+    }
+
+    @DisplayName("스케줄을 생성한다")
+    @Test
+    public void createSchedule() {
+        ScheduleRequest body = new ScheduleRequest(themeId, "2022-08-11", "13:00");
+        RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(body)
+                .when().post("/schedules")
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value());
     }
 
-    @DisplayName("관리자 외에는 테마를 생성할 수 없다")
+    @DisplayName("스케줄을 조회한다")
     @Test
-    public void createByUser() {
-        ThemeRequest body = new ThemeRequest("테마이름", "테마설명", 22000);
-        RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(body)
-                .when().post("/admin/themes")
-                .then().log().all()
-                .statusCode(HttpStatus.UNAUTHORIZED.value());
-    }
-
-    @DisplayName("관리자는 테마를 삭제할 수 있다")
-    @Test
-    void deleteByAdmin() {
-        Long id = createTheme();
+    public void showSchedules() {
+        requestCreateSchedule();
 
         var response = RestAssured
                 .given().log().all()
-                .header("Authorization", this.adminAccessToken)
-                .when().delete("/admin/themes/" + id)
+                .param("themeId", themeId)
+                .param("date", "2022-08-11")
+                .when().get("/schedules")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+
+        assertThat(response.jsonPath().getList(".").size()).isEqualTo(1);
+    }
+
+    @DisplayName("예약을 삭제한다")
+    @Test
+    void delete() {
+        String location = requestCreateSchedule();
+
+        var response = RestAssured
+                .given().log().all()
+                .when().delete(location)
                 .then().log().all()
                 .extract();
 
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
     }
 
-    @DisplayName("관리자 외에는 테마를 삭제할 수 없다")
-    @Test
-    void deleteByUser() {
-        Long id = createTheme();
-
-        var response = RestAssured
+    public static String requestCreateSchedule() {
+        ScheduleRequest body = new ScheduleRequest(1L, "2022-08-11", "13:00");
+        return RestAssured
                 .given().log().all()
-                .when().delete("/admin/themes/" + id)
-                .then().log().all()
-                .extract();
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
-    }
-
-    public Long createTheme() {
-        ThemeRequest body = new ThemeRequest("테마이름", "테마설명", 22000);
-        String location = RestAssured
-                .given().log().all()
-                .header("Authorization", this.adminAccessToken)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(body)
-                .when().post("/admin/themes")
+                .when().post("/schedules")
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value())
-                .extract().header("Location");
-        return Long.parseLong(location.split("/")[2]);
+                .extract()
+                .header("Location");
     }
 }
