+++ src/test/java/nextstep/dao/ThemeDaoTest.java
@@ -1,6 +1,5 @@
-package nextstep.theme;
+package nextstep.dao;
 
-import nextstep.dao.ThemeDao;
 import nextstep.domain.Theme;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
